+++ msort.h
@@ -1,7 +1,7 @@
 //! @file msort.h
 #ifndef MSORT_H_INCLUDED
 #define MSORT_H_INCLUDED
-void msort( void* m, int elcnt, int elsize, int (*cmp)(const void*, const void*));
+void msort( void* m, size_t elcnt, unsigned int elsize, int (*cmp)(const void*, const void*));
 
 
 #endif // MSORT_H_INCLUDED
